MODPUBSUB-212: Replace Future<Boolean> by Future<Void> in SecurityManager
Future<Boolean> is used with true indicating success and false indicating failure. This results in two ways how a failure can be propagated: As a failing Future, or as a succeeding Future returning false. This results in bloated code and is error-prone.

The idiomatic way is to only use the built-in success/failure indicator of the Future and to use Void as the type: Future<Void>.

This reveals that SecurityManagerImpl#addPermissions and #assignPermissions don't properly propagate failures.
ModTenantApiTest#shouldLoginWithEnvVarCredentials was adopted accordingly.

Switching from Future<Boolean> to Future<Void> also allowed to merge SecurityManager#loginPubSubUser and #getJWTToken.

Lessons learned:
Replacing Future<Boolean> by Future<Void> results in less code and uncovers bugs.