MODPUBSUB-213: Replace Future<Boolean> by Future<Void> in MessagingModuleDao
Future<Boolean> is used with true indicating success and false indicating failure. This results in two ways how a failure can be propagated: As a failing Future, or as a succeeding Future returning false. This results in bloated code and is error-prone.

The idiomatic way is to only use the built-in success/failure indicator of the Future and to use Void as the type: Future<Void>.

Approach:

Replace Future<Boolean> by Future<Void> and use the built-in error handling of Vert.x' Future: https://github.com/folio-org/raml-module-builder/blob/master/doc/futurisation.md#result-and-error-handling

MessagingModuleDao#delete(String id) is removed because it is not used.

The Boolean was set to "rowCount() == 1", however, the Boolean was not checked, a false Boolean was considered a successful delete.
Therefore the check is removed.